+++ CHANGELOG.md
@@ -11,10 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added version to all plugins starting at 0.1.0 - [ripe-white/#1026](https://github.com/ripe-tech/ripe-white/issues/1026)
 * Added html support for ´alert´ component  - [ripe-white/#1037](https://github.com/ripe-tech/ripe-white/issues/1037)
+* Support thumbnails for frames that are for personalization only (contain `group` in the thumbnails spec) - [ripe-white/#978](https://github.com/ripe-tech/ripe-white/issues/978)
 
 ### Changed
 
-*
+* Support for `showInitials` in `initials-images` as prop - [ripe-white/#978](https://github.com/ripe-tech/ripe-white/issues/978)
 
 ### Fixed